+++ Bermuda/Bermuda/Camera.h
@@ -2,16 +2,16 @@
 class Camera
 {
 private:
-	int x;
-	int y;
-	int width;
-	int height;
+	double x;
+	double y;
+	double width;
+	double height;
 public:
-	Camera(int x, int y, int width, int height);
-	int getX();
-	int getY();
-	int getWidth();
-	int getHeight();
+	Camera(double x, double y, double width, double height);
+	double getX();
+	double getY();
+	double getWidth();
+	double getHeight();
 	~Camera();
 };
